LayoutTests/media/track/track-webvtt-tc012-outoforder.html

<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            
 
            function trackLoaded()
            {
                findMediaElement();
                testExpected("video.textTracks[0].cues.length", "3");
                
                endTest();
            }

        </script>
    </head>
    <body>
    <p>Tests cues that are temporally out of order (we allow this).</p>
        <video>
            <track src="captions/tc012_outoforder.vtt" onload="trackLoaded()">
        </video>
    </body>
</html>
